Fixed shop not updating boot ability display on tab change
The shop's burst ability window assumed that boots have no abilities, this fix simply does away with this boot check.

The check probably exists in other places as well, but in this menu my new code happened to not be before the check, making it be skipped in the case of boots.