Fixed judge reward when "Judge PTS?" is off
JP is granted in a different place when the option "Judge PTS?" is off, I was not aware of this and it was causing the reward to not be updated (it was vanilla instead of using the reward definition) and, even worse, it was causing the JP to be capped at 10, often doing away with most of a character's JP when they got rewarded.

The solution is very similar to the original judgefix, only on a different spot.