Customizable name entry screen portraits
This allows to change which units the game will look for when getting the portraits for the main character and clan naming screens.

The units need to exist in ram for this to work, so they need to be either loaded through events or be part of the main party.

If using quickStart, the necessary characters will be loaded through the inserted event automatically, but if quickStart is not being used a custom event is necessary.