4 digits HP and MP display option
This is activated by morePointDigits.

While it is installed the unit stat menu will only show max HP and MP, and the smaller in-battle HUD (for targets and such) will only display current HP and MP.
The bigger HUD (the one that includes a portrait) still shows both current and max.

Note, however, that having this display means nothing if the newLevelUp option is not defined, as the HP and MP stats are capped at 999 in vanilla.

Additionally, now the newLevelUp option's newMaxStat and newMaxPoints (a new option to set a different limit for HP/MP) will apply to level ups even when levelUpUnaffected is defined.