Added a few small fixes
Exorcised units are no longer permanently deleted. This is extremely rare to happen in vanilla and is never explained, as well as being slightly glitchy.

Beneficial status effects no longer get 0% accuracy when the unit has a very large status defense. Having the highest bit of status defense set will still prevent beneficial status effects, however, as that is a flag marking the job immune to all status effects.

Prevented the game from reading the job... mask? whatever that thing is, if 0xFF it makes a job have the stats of the unit's first action command ID's job, so it's essentially useless. In vanilla it's used by Marche and other units for no good reason, there are already changes to prevent marche from having the special job anyway so this byte is useless and could only lead to confusion.

Prevented the game from setting an unused byte (0x0B) in unit data, essentially freeing it for future use. This byte seems to have originally been related to elemental stuff, but does absolutely nothing in vanilla as far as I can tell.

And finally, cleaned up a bit the defines and ifdefs in jpLearn.event.