Fixed item effect stacking bug
TLDR: Game didn't use enough bits to track applied effects, I made it use more.

The problem:
In vanilla, the game tries to keep track of item effects that have already been applied so that they won't be applied twice from different items.
However, it only allocates 16 bits to this, making any effect past that be able to stack.

Since most item effects do nothing extra when stacked ("null: element", "null: status" and such) this would have been fine as long as they had the "enhance: element" effects be the first few of the list... Which they did not.

As a result, the first few "enhance" effects do not stack (fire, wind, earth), while the rest do stack. I can't help but wonder if they moved them around at some point and they stopped working, or if they only tested if the fire one was working or something. Anyway...

The solution:
This engine hacks allocates 0x20 bytes of the stack to be used as a bitfield of already applied item effects, making the game able to keep track and prevent the stacking of up to 0xFF item effects. This was clearly the intended behavior so it is the default solution. However, an alternate solution is provided, by defining itemEffectStackAlways the whole stacking check will be skipped, making all effects able to stack.