inventory rework (probably broke something)
Long story short: In vanilla, items in the inventory take up 4 bytes in RAM, this makes them take only one, making it possible to expand the item list without fear of having inventory issues. The extra saved space will be used for... something I guess, in the future.

A bunch of new options related to the stuff that was changed have also been added, check options.event and data.event for more info.

Also added the noStatVariance to the definitions in the main file and improved the padding at the end of the ROM.